wip_srv_dev/requirements.txt
pymongo
python-dotenv
flask
bson
openai
beautifulsoup4
lxml
mysql-connector-python
janome
mecab-python3
unidic-lite
